+++ src/components/Sort/Sort.scss
@@ -1,0 +1,36 @@
+.sort {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 1.5rem;
+  margin-top: 1rem;
+
+  &__buttons {
+    display: flex;
+    gap: 1rem;
+  }
+
+  &__button {
+    display: flex;
+    gap: 1rem;
+    width: 180px;
+    font-size: 1rem;
+    padding: 0.5rem;
+  }
+
+  &__img {
+    width: 1.5rem;
+    height: 1.5rem;
+
+    &.hidden {
+      display: none;
+    }
+  }
+}
+
+@media (max-width: 419.98px) {
+  .sort {
+    flex-direction: column;
+    gap: 1rem;
+  }
+}
