+++ src/components/VehiclesList/VehiclesList.jsx
@@ -2,12 +2,10 @@ import React, { useState, useEffect } from 'react';
 
 import './VehiclesList.scss';
 
-import ascending from '../../assets/ascending.svg';
-import descending from '../../assets/descending.svg';
-
 import Spinner from '../Spinner/Spinner';
 import Error from '../Error/Error';
 import VehicleCard from '../VehicleCard/VehicleCard';
+import Sort from '../Sort/Sort';
 
 import VehicleService from '../../service/VehicleService';
 
@@ -46,22 +44,6 @@ const VehiclesList = () => {
     setLoading(false);
   };
 
-  const handleSortChange = (type) => {
-    if (type === sortType) {
-      setSortOrder((oldOrder) => oldOrder * -1);
-    } else {
-      if (type === 'year') {
-        setSortType('year');
-      }
-
-      if (type === 'price') {
-        setSortType('price');
-      }
-
-      setSortOrder(1);
-    }
-  };
-
   const renderVehiclesList = (vehiclesData) => {
     const vehiclesItems = vehiclesData.map((vehicleData) => {
       const { id } = vehicleData;
@@ -83,32 +65,12 @@ const VehiclesList = () => {
       <div className="vehicles-list container">
         <h2 className="vehicles-list__header">Список автомобилей</h2>
 
-        <div className="vehicles-list__sort">
-          <h3 className="vehicles-list__sort-title">Сортировка:</h3>
-
-          <button
-            className="vehicles-list__sort-button"
-            onClick={() => handleSortChange('year')}
-          >
-            <span>По году выпуска</span>
-            <img
-              className={`vehicles-list__sort-img ${sortType === 'year' ? '' : 'hidden'}`}
-              src={sortOrder > 0 ? ascending : descending}
-              alt={sortOrder > 0 ? 'ascending' : 'descending'}
-            />
-          </button>
-          <button
-            className="vehicles-list__sort-button"
-            onClick={() => handleSortChange('price')}
-          >
-            <span>По цене</span>
-            <img
-              className={`vehicles-list__sort-img ${sortType === 'price' ? '' : 'hidden'}`}
-              src={sortOrder > 0 ? ascending : descending}
-              alt={sortOrder > 0 ? 'ascending' : 'descending'}
-            />
-          </button>
-        </div>
+        <Sort
+          sortType={sortType}
+          setSortType={setSortType}
+          sortOrder={sortOrder}
+          setSortOrder={setSortOrder}
+        />
 
         {error}
         {spinner}
