+++ src/components/VehiclesList/VehiclesList.scss
@@ -15,36 +15,6 @@
     margin-top: 1rem;
     list-style: none;
   }
-
-  &__sort {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 1.5rem;
-    margin-top: 1rem;
-
-    &-buttons {
-      display: flex;
-      gap: 1rem;
-    }
-
-    &-button {
-      display: flex;
-      gap: 1rem;
-      width: 180px;
-      font-size: 1rem;
-      padding: 0.5rem;
-    }
-
-    &-img {
-      width: 1.5rem;
-      height: 1.5rem;
-
-      &.hidden {
-        display: none;
-      }
-    }
-  }
 }
 
 @media (max-width: 991.98px) {
@@ -60,11 +30,6 @@
     &__list {
       padding: 0 2rem;
       grid-template-columns: repeat(1, 1fr);
-    }
-
-    &__sort {
-      flex-direction: column;
-      gap: 1rem;
     }
   }
 }
